Reset the journey back to the submit to nrm question if the user changes whether they are under 18 or if they want a referral/dtn
